% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_data}
\alias{get_data}
\alias{get_surveys}
\alias{get_continents}
\alias{get_sources}
\alias{get_scientific_names}
\alias{get_catch_table}
\alias{get_taxon_table}
\alias{get_haul_table}
\alias{get_survey_table}
\title{Query FishGlob survey data}
\usage{
get_data(
  scientific_name = NULL,
  regex = FALSE,
  family = NULL,
  survey = NULL,
  continent = NULL,
  source = NULL,
  year = NULL
)

get_surveys()

get_continents()

get_sources()

get_scientific_names()

get_catch_table()

get_taxon_table()

get_haul_table()

get_survey_table()
}
\arguments{
\item{scientific_name}{Scientific name(s). See \code{\link[=get_scientific_names]{get_scientific_names()}}.}

\item{regex}{If \code{TRUE}, interpret \code{scientific_name} as
regular expressions. For example, \code{"^hippo*"} matches names starting with "hippo".}

\item{family}{Taxonomic family name(s). See \code{\link[=get_taxon_table]{get_taxon_table()}}.}

\item{survey}{Survey name(s). See \code{\link[=get_surveys]{get_surveys()}}.}

\item{continent}{Continent name(s), e.g. "europe". See \code{\link[=get_continents]{get_continents()}}.}

\item{source}{Data source name(s), e.g. "NOAA" or "DATRAS ICES". See \code{\link[=get_sources]{get_sources()}}.}

\item{year}{Vector of years. Defaults to all.}
}
\value{
A data frame of matching data.
}
\description{
Functions to query the FishGlob database. These functions assume the user has already run \code{\link[=install_fishglob_data]{install_fishglob_data()}} once to cache the FishGlob data in a local database on their computer.

\code{get_data()}: The main data querying function. Returns a data frame of matching survey catch records with joined haul, survey, and taxon information.

\code{get_surveys()}: Return available survey-unit IDs.

\code{get_continents()}: Return available continents.

\code{get_sources()}: Return available data sources.

\code{get_scientific_names()}: Return available accepted scientific names.

\code{get_catch_table()}: Return the complete catch table.

\code{get_taxon_table()}: Return the complete taxon table.

\code{get_haul_table()}: Return the complete haul table.

\code{get_survey_table()}: Return the complete survey table.
}
\details{
\code{get_data()} is the main querying function to return filtered survey catch records. Not all columns are included by default. Additional columns can be
appended by joining (e.g., \code{\link[dplyr:left_join]{dplyr::left_join()}}) the data returned by \code{get_data()} with the data from \code{get_taxon_table()}, \code{get_haul_table()}, or \code{get_survey_table()}.
}
\examples{
\dontrun{
# The first time, start by installing the
# database locally:
install_fishglob_data()

# Subsequently, the data does not need to be recached.

# Find available agency data sources
get_sources()

# The full list of available species:
spp <- get_scientific_names()

# All Canary Rockfish data held by DFO
dat <- get_data(
  scientific_name = "sebastes pinniger",
  source = "DFO"
)

# All Sebastes data globally from 2000 to 2021
dat <- get_data(
  scientific_name = "^sebastes*",
  regex = TRUE,
  year = 2000:2021
)
}

}
\seealso{
\code{\link[=install_fishglob_data]{install_fishglob_data()}}
}
