% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-data.R
\name{install_fishglob_data}
\alias{install_fishglob_data}
\title{Download and install FishGlob data}
\usage{
install_fishglob_data(force = FALSE)
}
\arguments{
\item{force}{If \code{TRUE}, download all tables again and rebuild the database.}
}
\value{
Invisibly, the path to the cached DuckDB database.
}
\description{
Downloads the FishGlob .rds tables from Dropbox into a cache folder and
builds a local DuckDB database. Later sessions reuse that database.
Intended to be used before \code{\link[=get_data]{get_data()}}.
}
\examples{
\donttest{
install_fishglob_data()
# then see ?get_getdata
}

}
\seealso{
\code{\link[=get_data]{get_data()}}
}
